Create README.md
Customer Churn refers to the rate of customer attrition in a company or we can say that spped at which customer laves your company or service.
For example,  cancellation of a subscription, closure of an account, decision to shop from another place, etc. .

Churn could happen due to many different reasons and churn analysis helps to identifi the cause ( and timings ) of this churn opening up opportunities to implement effective retention strategies .

Steps used for building a model to predict churn in a company are as follows :
  *   Gather available customer behaviour, transaction, demographics data and usage patterns.

 *    Utilize these data points to predict customer segments who are likely to churn.

 *     Create a model to pattern the risk tolerance of the business with respect to churn probability.

MODEL DESCRIPTION : 

The predictive churn model loos at the user activity from the past and checks who is active after certain.
Then creates a model that probabilistically identifies the steps and stages when a customer ( or segment ) is leaving your service.

Techniques used for building predictive models are:
1. Machine Learning
2. Deep Learning

MACHINE LEARNING : 
Machine learning relies on finding patterns and relationships in large amounts of data, the rulles discovered by the machuine learning model are guaranteed to be supported by evidence instead of intuition/ hunches.

Unlike human, who are limited by the number of variable/factors they can account for when crafting their business rules, Machine Learning algoriths can process and extract patterns from many variable.

Machine Learning is able to learn highly accurate rules in a much shorter time when compared to a human which usually needs a significant amount of experience and domain knowledge to devise accurate rules.

Machine Learning is able to timely detect concept drift and adapt the rules accordingly.

DEEP LEARNING : 
Deep learning is a subset of machine learning technique.

It teaches a computer or machine to filter inputs through layers to learn  how to predict and classify information.
Observations can be in the form of images, text, or sound.

Deep learning uses neural networks to simulate human-like decision making.
Deep learning models are capable to focus on the right features by themselves, requiring little guidance from the programmer.